+++ backend/BankingApplication/src/main/java/org/example/bankingapplication/dto/transaction/TransactionResponseDTO.java
@@ -6,6 +6,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.example.bankingapplication.enums.TransactionType;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -18,8 +19,8 @@ public class TransactionResponseDTO {
     private UUID transactionId;
     private String senderAccountNumber;
     private String recipientAccountNumber;
-    private Double amount;
-    private Double balance;
+    private BigDecimal amount;
+    private BigDecimal balance;
     private TransactionType transactionType;
     private LocalDateTime transactionDate;
     private String message;
